move CoursePane into RightPane put DesktopTabs in RightPane because it is the parent for map and added courses as well
