+++ src/components/RightPane/CoursePane/CoursePaneButtonRow.js
