+++ src/components/RightPane/CoursePane/SchoolDeptCard.js
